change position from usize to f64

+++ src/component.rs
@@ -39,35 +39,35 @@ pub trait PositionComponent {
     /// the node's position.
     /// # Returns
     /// * `usize` - The X coordinate of the top-left corner of the node.
-    fn get_x(&self) -> usize;
+    fn get_x(&self) -> f64;
 
     /// Get the Y position of the node.
     /// This function should return the Y coordinate of the top-left corner of
     /// the node's position.
     /// # Returns
     /// * `usize` - The Y coordinate of the top-left corner of the node.
-    fn get_y(&self) -> usize;
+    fn get_y(&self) -> f64;
 
     /// Set the X position of the node.
     /// The position is the top-left corner of the node, so the passed `x` value
     /// should be the X coordinate of the top-left corner.
     /// # Arguments
     /// * `x` - The new X coordinate of the top-left corner of the node
-    fn set_x(&mut self, x: usize);
+    fn set_x(&mut self, x: f64);
 
     /// Set the Y position of the node.
     /// The position is the top-left corner of the node, so the passed `y` value
     /// should be the Y coordinate of the top-left corner.
     /// # Arguments
     /// * `y` - The new Y coordinate of the top-left corner of the node
-    fn set_y(&mut self, y: usize);
+    fn set_y(&mut self, y: f64);
 
     /// Set the position of the node to the given `(x, y)` tuple.
     /// The position is the top-left corner of the node.
     /// # Arguments
     /// * `position` - A tuple containing the X and Y coordinates of the new
     ///   position, in the form (x, y).
-    fn set_position(&mut self, position: (usize, usize)) {
+    fn set_position(&mut self, position: (f64, f64)) {
         self.set_x(position.0);
         self.set_y(position.1);
     }
@@ -80,7 +80,7 @@ pub trait PositionComponent {
     /// # Returns
     /// * `(usize, usize)` - A tuple containing the X and Y coordinates of the
     ///   top-left corner of the node, in the form (x, y).
-    fn get_position(&self) -> (usize, usize) {
+    fn get_position(&self) -> (f64, f64) {
         (self.get_x(), self.get_y())
     }
 }
